V1.2

+++ apex/submit.py
@@ -75,10 +75,15 @@ def pack_upload_dir(
     copy_all_other_files(work_dir, upload_dir, ignore_list=ignore_copy_list)
     for ii in conf_dirs:
         build_conf_path = os.path.join(upload_dir, ii)
-        copy_poscar_path = os.path.abspath(os.path.join(ii, "POSCAR"))
-        target_poscar_path = os.path.join(build_conf_path, "POSCAR")
         os.makedirs(build_conf_path, exist_ok=True)
-        shutil.copy(copy_poscar_path, target_poscar_path)
+        copy_poscar_path = os.path.abspath(os.path.join(ii, "POSCAR"))
+        copy_stru_path = os.path.abspath(os.path.join(ii, "STRU"))
+        if os.path.isfile(copy_poscar_path):
+            target_poscar_path = os.path.join(build_conf_path, "POSCAR")
+            shutil.copy(copy_poscar_path, target_poscar_path)
+        if os.path.isfile(copy_stru_path):
+            target_stru_path = os.path.join(build_conf_path, "STRU")
+            shutil.copy(copy_stru_path, target_stru_path)
         if flow_type == 'props':
             copy_relaxation_path = os.path.abspath(os.path.join(ii, "relaxation"))
             target_relaxation_path = os.path.join(build_conf_path, "relaxation")
